prune more dependencies
