+++ .github/workflows/coverage.yaml
@@ -59,6 +59,7 @@ jobs:
           set -euxo pipefail
 
           yq --version || brew install yq
+          mint --version || brew install mint
 
           swift run fishy-joes install-toolchains --swiftly --wasm
