+++ Sources/FishyJoesExecute/FileTemplater.swift
@@ -1,8 +1,6 @@
 import FishyJoesConfig
 import Foundation
-import GenerationHelpers
 import swsh
-import Yams
 
 public struct FileTemplater {
     let config: FishyJoesConfig
