+++ Sources/FishyJoesExecute/Resources/bindings-template/workflows/__TEMPLATE_CAPS__-kotlin.yaml
@@ -49,6 +49,7 @@ jobs:
         shell: zsh {0}
         run: |
           yq --version || brew install yq
+          mint --version || brew install mint
 
       - name: Setup node
         uses: actions/setup-node@v3
