+++ integration-tests/TestAPI/.github/workflows/GENERATED-c-sharp.yaml
@@ -50,6 +50,7 @@ jobs:
         shell: zsh {0}
         run: |
           yq --version || brew install yq
+          mint --version || brew install mint
 
       - name: Setup node
         uses: actions/setup-node@v3
