+++ integration-tests/TestAPI/bindings/workflows/GENERATED-typescript.yaml
@@ -42,6 +42,7 @@ jobs:
         shell: zsh {0}
         run: |
           yq --version || brew install yq
+          mint --version || brew install mint
 
       - name: Setup node
         uses: actions/setup-node@v3
